requirements.txt
streamlit
sentence-transformers
python-pptx
kss
docx